feat: создать хранилища для User и Film
