+++ .gitignore
@@ -49,3 +49,4 @@ db/*
 /mvnw.cmd
 /.mvn/
 *.json
+*.xml
