+++ src/main/java/ru/yandex/practicum/filmorate/storage/UserStorage.java
@@ -1,0 +1,15 @@
+package ru.yandex.practicum.filmorate.storage;
+
+import ru.yandex.practicum.filmorate.model.User;
+
+import java.util.List;
+
+public interface UserStorage {
+	List<User> getUsers();
+
+	User addUser(User user);
+
+	User updateUser(User user);
+
+	boolean validate(User user);
+}
