+++ src/test/java/ru/yandex/practicum/filmorate/FilmValidationTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.Test;
 import ru.yandex.practicum.filmorate.controller.FilmController;
 import ru.yandex.practicum.filmorate.exception.EntityValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.storage.InMemoryFilmStorage;
 
 import java.time.LocalDate;
 
@@ -12,10 +13,11 @@ import static org.junit.jupiter.api.Assertions.*;
 
 public class FilmValidationTest {
 	FilmController filmController;
+	InMemoryFilmStorage inMemoryFilmStorage = new InMemoryFilmStorage();
 
 	@BeforeEach
 	void clearFilmController() {
-		filmController = new FilmController();
+		filmController = new FilmController(inMemoryFilmStorage);
 	}
 
 	@Test
