+++ src/test/java/ru/yandex/practicum/filmorate/UserValidationTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.Test;
 import ru.yandex.practicum.filmorate.controller.UserController;
 import ru.yandex.practicum.filmorate.exception.EntityValidationException;
 import ru.yandex.practicum.filmorate.model.User;
+import ru.yandex.practicum.filmorate.storage.InMemoryUserStorage;
 
 import java.time.LocalDate;
 
@@ -12,10 +13,11 @@ import static org.junit.jupiter.api.Assertions.*;
 
 public class UserValidationTest {
 	UserController userController;
+	InMemoryUserStorage inMemoryUserStorage = new InMemoryUserStorage();
 
 	@BeforeEach
 	void clearUserController() {
-		userController = new UserController();
+		userController = new UserController(inMemoryUserStorage);
 	}
 
 	@Test
